Opt out from .NET telemetry

+++ build.bat
@@ -1,10 +1,12 @@
 @echo off
 
+set DOTNET_CLI_TELEMETRY_OPTOUT=1
+
 for /F "tokens=*" %%f in ('dir /B /AD /S bin') do rmdir /S /Q "%%f"
 for /F "tokens=*" %%f in ('dir /B /AD /S obj') do rmdir /S /Q "%%f"
-rmdir /S /Q Build
-rmdir /S /Q Debug
-rmdir /S /Q Release
+if exist Build rmdir /S /Q Build
+if exist Debug rmdir /S /Q Debug
+if exist Release rmdir /S /Q Release
 
 mkdir Build
 mkdir Build\VNTextPatch
